assignment02/requirements.txt
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
tzdata==2025.1
xpress==9.4.2
xpresslibs==9.4.2
